Finally have the CalendarEntry component working!
It can be used from multiple places ...
... but I wish I could stay in list view and change the calendar. There
might be a way to do it the way I want, but it's not working for me.
I tried to put a key attribute on the list component, but it didn't make
a difference. It only works if I put a key attribute on the catch-all
child component in the main container ...

NOTE: I want to merge to master, rebase to remove non-sense, and then
push to master